requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
google-cloud-secret-manager
google-cloud-pubsub
psycopg2-binary
marker-pdf
